bugfix: type of argument in close session frontend

+++ client/components/CardSession.tsx
@@ -19,7 +19,7 @@ const MutationStartSession = `
 `;
 
 const MutationCloseSession = `
-    mutation ($id: Float!) {
+    mutation ($id: Int!) {
         closeSession(id: $id) {
             session {
                 endTime
